feat(list): add list component
Introduces the list component with two modifiers (inline and grid) that change the presentation of list items. If the list component is used without modifiers, it gets the same styling as lists within the content component.

The inline modifier displays the list items in a horizontal line using flexbox. The grid modifier positions list items within a flexible grid with a minimum column width of 100px.