fix(login): Add basic form and list components to demo (#785)
* fix(login): Add basic form components to demo

Adds form components to the login demo. Input styling doesn’t align with the mockup, but should in the future when the new FormControl is implemented.

* Fixed incorrect for attribute of labels

For attribute of labels didn’t match the ID of form elements.

* Fixed dropdown toggle icon

Updated handlebars syntax to make the toggle icon appear correctly.

* Removed alt text from decorative logo images

Decorative images like these don’t need to include alt text that’s more or less the same as the following <p> tags.

* Added links to list items

The SSO and footer items should have been links, and now they are. Note that the contrast of footer links against the background probably does not pass our requirements and will need to be fixed separately.

* Replace Dropdown with Select variation

Switched to use the new Select variation of the Dropdown component.

* Adjusted bottom margin of login box title

* Adjusted margins of form button and checkbox

Right margin of the form’s button is now 24px as shown in the mockup, and the bottom margin of the checkbox wasn’t necessary.

* Turned "Or log in with" text into an h2 title

* Simplified list item markup

Added display--element="a" to the display utility to remove the need for a parent a tag around each item.

* Hide social text labels on mobile and misc fixes

* Removed toolbar containers from form actions

An alternative (and possibly temporary) approach to structuring the action area of the form. Still TBD whether this approach with a button modifier or toolbar-group with other modifiers/utilities would be the best way to achieve the desired layout.

* Added margin to checkbox for mobile layout

* Added new width utility breakpoints to button

If we implement a new width utility as proposed, this demo’s Sign In button will resize correctly.

* Removed unnecessary aria-describedby attributes

The inputs don’t include helper text, so the attribute was not necessary.

* Changed “Sign in” to “Log in”

Missed it in one place.

* Use Flex utilities to achieve desired layout

Flexbox’s align-self:stretch and align-self:flex-start ended up being the key to achieve the desired responsive layout without the need for additional utilities or modifiers.

* Align checkbox vertically

* Removed footer link custom styling